Add support to both JSON and Table formats
