+++ cli/create_server_operation.go
@@ -63,7 +63,7 @@ func runOperationServersCreateServer(cmd *cobra.Command, args []string) error {
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
