+++ cli/create_server_reinstall_operation.go
@@ -66,7 +66,7 @@ func runOperationServerReinstallCreateServerReinstall(cmd *cobra.Command, args [
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
