+++ cli/delete_api_key_operation.go
@@ -61,7 +61,7 @@ func runOperationAPIKeysDeleteAPIKey(cmd *cobra.Command, args []string) error {
 	}
 
 	if !debug {
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
