+++ cli/delete_project_operation.go
@@ -61,7 +61,7 @@ func runOperationProjectsDeleteProject(cmd *cobra.Command, args []string) error 
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
