+++ cli/destroy_server_operation.go
@@ -61,7 +61,7 @@ func runOperationServersDestroyServer(cmd *cobra.Command, args []string) error {
 	}
 
 	if !debug {
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
