+++ cli/get_api_keys_operation.go
@@ -59,7 +59,7 @@ func runOperationAPIKeysGetAPIKeys(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	if !debug {
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
