+++ cli/get_bandwidth_plans_operation.go
@@ -63,7 +63,7 @@ func runOperationPlansGetBandwidthPlans(cmd *cobra.Command, args []string) error
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
