+++ cli/get_plan_operation.go
@@ -62,7 +62,7 @@ func runOperationPlansGetPlan(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	if !debug {
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
