+++ cli/get_plans_operation.go
@@ -95,7 +95,7 @@ func runOperationPlansGetPlans(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	if !debug {
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
