+++ cli/get_project_operation.go
@@ -67,7 +67,7 @@ func runOperationProjectsGetProject(cmd *cobra.Command, args []string) error {
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
