+++ cli/get_projects_operation.go
@@ -78,7 +78,7 @@ func runOperationProjectsGetProjects(cmd *cobra.Command, args []string) error {
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
