+++ cli/get_servers_operation.go
@@ -111,7 +111,7 @@ func runOperationServersGetServers(cmd *cobra.Command, args []string) error {
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
