+++ cli/post_project_ssh_key_operation.go
@@ -67,7 +67,7 @@ func runOperationSSHKeysPostProjectSSHKey(cmd *cobra.Command, args []string) err
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
