+++ cli/update_api_key_operation.go
@@ -66,7 +66,7 @@ func runOperationAPIKeysUpdateAPIKey(cmd *cobra.Command, args []string) error {
 		return err
 	}
 	if !debug {
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
