+++ cli/update_server_operation.go
@@ -66,7 +66,7 @@ func runOperationServersUpdateServer(cmd *cobra.Command, args []string) error {
 	}
 	if !debug {
 
-		utils.PrintResult(msgStr)
+		utils.PrintResult(msgStr, "table")
 	}
 	return nil
 }
